Improve resource merging error handling. DO NOT MERGE
This changeset tracks changes in sdk-common to throw MergingExceptions when there
is a problem with resoure merging (XML errors, duplicate errors, file access errors,
invalid file name errors, and so on). It catches these and emits the errors (which
provide a format easily parseable in IDEs, always including the filename, line
number, and column if available, etc).

Change-Id: I25a41323bea085036b6c1ca5b533d5b6c221dbb6